CHROMIUM: soc: mediatek: SCPSYS: Drop parent/child relationship for mfg domains
The mfg power domains do not have a real parent/child relationship, so
the current implementation does not reflect reality. Nevertheless the
different mfg domains have hardware constraints in that they must be
enabled and disabled in a certain order.
To make that sure we no longer register the mfg_2d and mfg_async as
domains themselves but instead only the mfg domain which then also
enables/disabled the other two domains.

Note: This is a non-upstreamable CHROMIUM patch, because while it is
assumed that this is actually a work-around for an as-yet undiagnosed
power sequencing issue in the drm/img-rogue GPU driver.
An upstream solution would allow a fully functioning GPU driver to
also operate in the following "intermediate" power states:
 (a) mfg_async
 (b) mfg_async + mfg_2d

This fix in this patch disables these two modes since it always
enables/disables all three power domains together, mfg_async/mfg_2d/mfg.

For chromeos-4.4, this is a squash of:
4762f11e0c95 CHROMIUM: soc: mediatek: SCPSYS: Drop parent/child relationship for mfg domains
537dccbc8eee BACKPORT: soc: mediatek: SCPSYS: Fix double enabling of regulators
             CHROMIUM: soc: mediatek: SCPSYS: Fix power off sequence in error path
 (https://chromium-review.googlesource.com/#/c/448737/)

The BACKPORT was actually reverted upstream, since it really just
fixes an implementation detail of the CHROMIUM patch - mainline
does not have mfg domains, and hence doesn't have a 'double enabling issue'.

The BACKPORT's commit message:

With CONFIG_PM enabled do not call genpd->power_on manually as this
will cause the regulators being turned on once in SCPSYS probe and
then again when the genpd core turns on the domains. Instead, call
genpd->power_on only with CONFIG_PM disabled and tell the genpd core
that the domains are disabled when registered.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>

BUG=chrome-os-partner:45713
TEST=build and boot oak-rev4
TEST=Let system idle over night
 => No messages like:
    "mtk-scpsys 10006000.scpsys: Failed to power off domain mfg"
TEST=Run power_LoadTest
 => No messages like:
    "mtk-scpsys 10006000.scpsys: Failed to power off domain mfg"
BUG=chrome-os-partner:49416
TEST=On oak-rev4:
    stop ui
    localhost ~ # cat /sys/kernel/debug/pm_genpd/pm_genpd_summary
    domain                          status          slaves
        /device                                             runtime status
        ----------------------------------------------------------------------
        mfg                             off
    localhost ~ # cat /sys/class/regulator/regulator.10/name
    vgpu
    localhost ~ # cat /sys/class/regulator/regulator.10/status
    off

Reviewed-on: https://chromium-review.googlesource.com/317710
Reviewed-on: https://chromium-review.googlesource.com/322044
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>

Change-Id: Id53ef23b0ec215a57c7ae1c9d17dc89b63baea33
Reviewed-on: https://chromium-review.googlesource.com/423591
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>