ANDROID: sched/fair: filter energy_diff() based on energy_payoff value
Once the SchedTune support is enabled and the CPU bandwidth demand of a
task is boosted, we could expect increased energy consumptions which are
balanced by corresponding increases of tasks performance.
However, the current implementation of the energy_diff() function
accepts all and _only_ the schedule candidates which results into a
reduced expected system energy, which works against the boosting
strategy.

This patch links the energy_diff() function with the "energy payoff"
engine provided by SchedTune. The energy variation computed by the
energy_diff() function is now filtered using the SchedTune support to
evaluated the energy payoff for a boosted task.

With that patch, the energy_diff() function is going to reported as
"acceptable schedule candidate" only the schedule candidate which
corresponds to a positive energy_payoff.

BUG=chrome-os-partner:58667
TEST=build/boot on kevin

Change-Id: I1d86cad2fe790a42fab0d16cab4aa2ae889a5d08
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
(cherry picked from commit a2a6dc750833243b3b49d016291ae1133196759
 from android-msm/android-4.4.y)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/420037
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>