UPSTREAM: component: Detach components when deleting master struct
component_master_add_with_match calls find_components which, if any
components already exist, it attaches to the master struct. However, if
we later encounter an error the master struct is deleted, leaving
components with a dangling pointer to it.

If the error was a temporary one, e.g. for probe deferral, then when
the master device is re-probed, it will fail to find the required
components as they appear to already be attached to a master.

Fix this by nulling components pointers to the master struct when it is
deleted. This code is factored out into a separate function so it can be
shared with component_master_del.

Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
(cherry picked from commit 57480484f9f7631738ef28b952eca3c9081c4291)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:60805
TEST=build chromeos-4.4 for elm and boot
TEST=build chromeos-4.4 for kevin and boot

BUG=chrome-os-partner:50168
TEST=With entire series applied,
TEST=1) stop cras
TEST=2) unload machine driver
TEST=3) echo 0000:00:1f.3 > \
TEST=   /sys/module/snd_soc_skl/drivers/pci:snd_soc_skl/unbind
TEST=kernel audio driver does not crash

Reviewed-on: https://chromium-review.googlesource.com/333125
Commit-Ready: Vedang Patel <vedang.patel@intel.com>
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>

Change-Id: I3695baf19a3f246298559669fb37963f77663845
Reviewed-on: https://chromium-review.googlesource.com/423755
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>