UPSTREAM: drm/i915: Remove the identical implementations of request space reservation
Now that we share intel_ring_begin(), reserving space for the tail of
the request is identical between legacy/execlists and so the tautology
can be removed. In the process, we move the reserved space tracking
from the ringbuffer on to the request. This is to enable us to reorder
the reserved space allocation in the next patch.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1461833819-3991-13-git-send-email-chris@chris-wilson.co.uk
(cherry picked from commit 0251a963226de25978920544c6e1f4878d544835
from drm-intel-next-2016-05-08)
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

BUG=chrome-os-partner:58799
TEST=Browse-able and bootable image
TEST=No other obvious regressions

Change-Id: I0add253b76bc688c6f9870b56deaa150e149842c
Reviewed-on: https://chromium-review.googlesource.com/401555
Commit-Ready: Wayne Boyer <wayne.boyer@intel.com>
Tested-by: Michael Cheng <michael.cheng@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>