FROMLIST: mmc: core: switch to 1V8 or 1V2 for hs400es mode
When introducing hs400es, I didn't notice that we haven't
switched voltage to 1V2 or 1V8 for it. That happens to work
as the first controller claiming to support hs400es, arasan(5.1),
which is designed to only support 1V8. So the voltage is fixed to 1V8.
But it actually is wrong, and will not fit for other host controllers.
Let's fix it.

BUG=chrome-os-partner:57924
TEST=boot on kevin & suspend_stress_test

(am from https://patchwork.kernel.org/patch/9347203/)
Fixes: commit 81ac2af65793ecf ("mmc: core: implement enhanced strobe support")

Change-Id: If074bfea701a9527e3a4a8c1fe7f5e109fd43156
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Signed-off-by: Ziyuan Xu <xzy.xu@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/390532
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Shawn Lin <shawn.rin@gmail.com>